requirements.txt
asgiref==3.11.0
attrs==25.4.0
Django==3.2.16
djangorestframework==3.12.4
drf-spectacular==0.29.0
drf-spectacular-sidecar==2026.1.1
exceptiongroup==1.3.1
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==26.0
Pillow==9.3.0
pluggy==1.6.0
Pygments==2.19.2
pytest==7.4.4
pytest-django==4.11.1
pytest-lazy-fixture==0.6.3
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.5
tomli==2.4.0
typing_extensions==4.15.0
uritemplate==4.2.0
